requirements.txt
# YEKA MedDikte — Bağımlılıklar (Mac + Linux)

faster-whisper>=1.0.0
flask>=3.0
flask-socketio>=5.3
flask-cors>=4.0
reportlab>=4.0
python-dotenv>=1.0
requests>=2.31
